Simplify and fix SQL types

+++ setup.sql
@@ -1,31 +1,37 @@
-drop table
+drop table if exists
   eat_sessions;
+
+drop function if exists
+	get_vote_count;
+
+drop function if exists
+	get_winner;
 
 create table
   public.eat_sessions (
     id uuid not null default uuid_generate_v4 (),
     created_at timestamp with time zone not null default now(),
-    options jsonb[] not null default '{}'::jsonb[],
-    votes jsonb[] not null default '{}'::jsonb[],
-    votes_required smallint not null default 2,
-    winner smallint null,
+    options jsonb not null default '[]'::jsonb,
+    votes jsonb not null default '[]'::jsonb,
+    votes_required integer not null default 2,
+    winner integer null,
     constraint eat_sessions_pkey primary key (id)
   ) tablespace pg_default;
 
 alter table
   eat_sessions enable row level security;
 
-create or replace function
+create function
 	get_vote_count(id uuid)
-		returns smallint
+		returns integer
 		language sql
 		as $$
-			select cardinality(votes)
+			select jsonb_array_length(votes)
 			from eat_sessions
 			where id = eat_sessions.id;
 		$$;
 
-create or replace function
+create function
 	get_winner(id uuid)
 		returns jsonb
 		language sql
